cambio parsefloat por parseint

+++ js/main.js
@@ -4,9 +4,9 @@ function calcularCostoCombustible() {
     let nuevoCalculo = true;
 
     do {
-        let distanciaViaje = parseFloat(prompt("Ingresá la distancia de tu viaje (en kilómetros, utiliza solo números)"));
-        let precioCombustible = parseFloat(prompt("Ingresá el precio del litro de combustible (sin el signo $, solo números)"));
-        let rendimientoAuto = parseFloat(prompt("Ingresá el rendimiento de tu auto por litro (en kilómetros por litro, utiliza solo números)"));
+        let distanciaViaje = parseInt(prompt("Ingresá la distancia de tu viaje (en kilómetros, utiliza solo números)"));
+        let precioCombustible = parseInt(prompt("Ingresá el precio del litro de combustible (sin el signo $, solo números)"));
+        let rendimientoAuto = parseInt(prompt("Ingresá el rendimiento de tu auto por litro (en kilómetros por litro, utiliza solo números)"));
 
         if (!isNaN(distanciaViaje) && !isNaN(precioCombustible) && !isNaN(rendimientoAuto)) {
             const litrosNecesarios = distanciaViaje / rendimientoAuto;
